Improve JavaDoc
