+++ src/main/java/command/AddCommand.java
@@ -5,10 +5,18 @@ import task.Task;
 import task.TaskList;
 import ui.Ui;
 
+/**
+ * Command for adding a task to a task list.
+ */
 public class AddCommand extends Command {
     private final char taskType;
     private final String content;
 
+    /**
+     * Constructor for AddCommand.
+     * @param taskType The type of task to add to the task list.
+     * @param commandContent The content to be added to the task.
+     */
     public AddCommand(char taskType, String commandContent) {
         this.taskType = taskType;
         this.content = commandContent;
