--- src/main/java/dukeException/CommandException/EmptyCommandException.java
@@ -1,7 +1,0 @@
-package dukeException.CommandException;
-
-public class EmptyCommandException extends UnknownCommandException {
-    public EmptyCommandException() {
-        super();
-    }
-}
