--- src/main/java/dukeException/CommandException/UnknownCommandException.java
@@ -1,9 +1,0 @@
-package dukeException.CommandException;
-
-import dukeException.DukeException;
-
-public class UnknownCommandException extends DukeException {
-    public UnknownCommandException() {
-        super("HUH? What you say?", null);
-    }
-}
