--- src/main/java/dukeException/DukeException.java
@@ -1,7 +1,0 @@
-package dukeException;
-
-public class DukeException extends RuntimeException {
-    public DukeException(String errorMessage, Throwable err) {
-        super(errorMessage, err);
-    }
-}
