--- src/main/java/dukeException/NotFoundException.java
@@ -1,7 +1,0 @@
-package dukeException;
-
-public class NotFoundException extends DukeException {
-    public NotFoundException(String source, String reason, Throwable err) {
-        super(String.format("Haiya nothing here in the %s lah. \n%s", source, reason), err);
-    }
-}
