--- src/main/java/dukeException/StorageException/CorruptedFileException.java
@@ -1,7 +1,0 @@
-package dukeException.StorageException;
-
-public class CorruptedFileException extends StorageException {
-    public CorruptedFileException() {
-        super("File Corrupted. FAILURE!");
-    }
-}
