--- src/main/java/dukeException/StorageException/CreateFileException.java
@@ -1,7 +1,0 @@
-package dukeException.StorageException;
-
-public class CreateFileException extends StorageException {
-    public CreateFileException() {
-        super("Create file FAILURE!");
-    }
-}
