--- src/main/java/dukeException/StorageException/LoadFileException.java
@@ -1,7 +1,0 @@
-package dukeException.StorageException;
-
-public class LoadFileException extends StorageException {
-    public LoadFileException() {
-        super("Load file FAILURE!");
-    }
-}
