--- src/main/java/dukeException/StorageException/SaveFileException.java
@@ -1,7 +1,0 @@
-package dukeException.StorageException;
-
-public class SaveFileException extends StorageException {
-    public SaveFileException() {
-        super("Save file FAILURE!");
-    }
-}
