--- src/main/java/dukeException/StorageException/StorageException.java
@@ -1,9 +1,0 @@
-package dukeException.StorageException;
-
-import dukeException.DukeException;
-
-public class StorageException extends DukeException {
-    public StorageException(String message) {
-        super(message, null);
-    }
-}
