+++ src/main/java/dukeexception/commandexception/EmptyCommandException.java
@@ -1,0 +1,13 @@
+package dukeexception.commandexception;
+
+/**
+ * Exception for handling an empty command.
+ */
+public class EmptyCommandException extends UnknownCommandException {
+    /**
+     * Constructor for EmptyCommandException.
+     */
+    public EmptyCommandException() {
+        super();
+    }
+}
