+++ src/main/java/dukeexception/commandexception/UnknownCommandException.java
@@ -1,0 +1,15 @@
+package dukeexception.commandexception;
+
+import dukeexception.DukeException;
+
+/**
+ * Exception for handling an unknown command.
+ */
+public class UnknownCommandException extends DukeException {
+    /**
+     * Constructor for UnknownCommandException.
+     */
+    public UnknownCommandException() {
+        super("HUH? What you say?", null);
+    }
+}
