+++ src/main/java/dukeexception/storageexception/CorruptedFileException.java
@@ -1,0 +1,13 @@
+package dukeexception.storageexception;
+
+/**
+ * Exception for handling failed file parsing.
+ */
+public class CorruptedFileException extends StorageException {
+    /**
+     * Constructor for CorruptedFileException.
+     */
+    public CorruptedFileException() {
+        super("File Corrupted. FAILURE!");
+    }
+}
