+++ src/main/java/dukeexception/storageexception/LoadFileException.java
@@ -1,0 +1,13 @@
+package dukeexception.storageexception;
+
+/**
+ * Exception for handling failed file loads.
+ */
+public class LoadFileException extends StorageException {
+    /**
+     * Constructor for LoadFileException.
+     */
+    public LoadFileException() {
+        super("Load file FAILURE!");
+    }
+}
