+++ src/main/java/dukeexception/storageexception/SaveFileException.java
@@ -1,0 +1,13 @@
+package dukeexception.storageexception;
+
+/**
+ * Exception for handling failed file saves.
+ */
+public class SaveFileException extends StorageException {
+    /**
+     * Constructor for SaveFileException.
+     */
+    public SaveFileException() {
+        super("Save file FAILURE!");
+    }
+}
