+++ src/main/java/struct/Triple.java
@@ -1,10 +1,22 @@
 package struct;
 
+/**
+ * Triple struct for handling complex data.
+ * @param <A> Type of first value.
+ * @param <B> Type of second value.
+ * @param <C> Type of third value.
+ */
 public final class Triple<A, B, C> {
     private final A a;
     private final B b;
     private final C c;
 
+    /**
+     * Constructor for Triple.
+     * @param a The first value.
+     * @param b The second value.
+     * @param c The third value.
+     */
     public Triple(A a, B b, C c) {
         this.a = a;
         this.b = b;
